Revert "docs: modify and add new squad data"
