+++ website/contents/data/quarter-leaders.js
@@ -4,167 +4,167 @@ export const leaders = [
     name: "Zi Ching Goh",
     position: "Subang Leader",
     linkedin: "https://www.linkedin.com/in/zi-ching-g-750b26198/",
-    image: "../../static/img/squad-profile-images/zi-ching-goh.jpg",
+    image: "https://media-exp1.licdn.com/dms/image/C5103AQEFXWZ97GxaNg/profile-displayphoto-shrink_400_400/0/1574999230863?e=1617235200&v=beta&t=axsNPLoEvA5xeYV0qqqw_GXD1uZw8Oi4BpnjgsNlqWU",
   },
   {
     quarter: "January - March 2021",
     name: "Hira Khalique",
     position: "Nilai Leader",
     linkedin: "https://www.linkedin.com/in/hira-khalique-9871211a3/",
-    image: "../../static/img/squad-profile-images/hira-khalique.jpg",
+    image: "https://media-exp1.licdn.com/dms/image/C5603AQG0FeISTDq8Ww/profile-displayphoto-shrink_400_400/0/1601366509956?e=1617235200&v=beta&t=1AvLcUYau37zfrUX4MgdY2JzB67D3Z70fN7ELbZXG78",
   },
   {
     quarter: "January - March 2021",
     name: "Jin En Ng",
     position: "Subang Vice Leader",
     linkedin: "https://www.linkedin.com/in/jin-en-ng-a343b51a3/",
-    image: "../../static/img/squad-profile-images/jin-en-ng.jpg",
+    image: "https://media-exp1.licdn.com/dms/image/C5103AQGZcxEEp7aKYg/profile-displayphoto-shrink_400_400/0/1582458683616?e=1617235200&v=beta&t=iX3x0h_ID5yAkXiXvP5yYV9WAQDSn3j-EfYuJuecWKI",
   },
   {
     quarter: "January - March 2021",
     name: "Krena Shree",
     position: "Nilai Vice Leader",
     linkedin: "https://www.linkedin.com/in/krena-shree-0953b61a3/",
-    image: "../../static/img/squad-profile-images/krena-shree.jpg",
+    image: "https://media-exp1.licdn.com/dms/image/C5603AQHJbZm85ItvdQ/profile-displayphoto-shrink_400_400/0/1611498868534?e=1617235200&v=beta&t=AXegKmg6SwkH3OUSo8kcBitrBy7bnwXH5sE6qSGSjUQ",
   },
   {
     quarter: "October - December 2020",
     name: "Isabella Chee",
     position: "Subang Leader",
     linkedin: "https://www.linkedin.com/in/isabella-chee-3142791b6/",
-    image: "../../static/img/squad-profile-images/isabella-chee.png",
+    image: "",
   },
   {
     quarter: "October - December 2020",
     name: "Hira Khalique",
     position: "Nilai Leader",
     linkedin: "https://www.linkedin.com/in/hira-khalique-9871211a3/",
-    image: "../../static/img/squad-profile-images/hira-khalique.jpg",
+    image: "https://media-exp1.licdn.com/dms/image/C5603AQG0FeISTDq8Ww/profile-displayphoto-shrink_400_400/0/1601366509956?e=1617235200&v=beta&t=1AvLcUYau37zfrUX4MgdY2JzB67D3Z70fN7ELbZXG78",
   },
   {
     quarter: "October - December 2020",
     name: "Zi Ching Goh",
     position: "Subang Vice Leader",
     linkedin: "https://www.linkedin.com/in/zi-ching-g-750b26198/",
-    image: "../../static/img/squad-profile-images/zi-ching-goh.jpg",
+    image: "https://media-exp1.licdn.com/dms/image/C5103AQEFXWZ97GxaNg/profile-displayphoto-shrink_400_400/0/1574999230863?e=1617235200&v=beta&t=axsNPLoEvA5xeYV0qqqw_GXD1uZw8Oi4BpnjgsNlqWU",
   },
   {
     quarter: "October - December 2020",
     name: "Krena Shree",
     position: "Nilai Vice Leader",
     linkedin: "https://www.linkedin.com/in/krena-shree-0953b61a3/",
-    image: "../../static/img/squad-profile-images/krena-shree.jpg",
+    image: "https://media-exp1.licdn.com/dms/image/C5603AQHJbZm85ItvdQ/profile-displayphoto-shrink_400_400/0/1611498868534?e=1617235200&v=beta&t=AXegKmg6SwkH3OUSo8kcBitrBy7bnwXH5sE6qSGSjUQ",
   },
   {
     quarter: "July - September 2020",
     name: "Zhen Hao Lee",
     position: "Subang Leader",
     linkedin: "https://www.linkedin.com/in/lee-zhen-hao-a28a96155/",
-    image: "../../static/img/squad-profile-images/zhen-hao-lee.jpg",
+    image: "https://media-exp1.licdn.com/dms/image/C5603AQEA5T3pKE6mFw/profile-displayphoto-shrink_400_400/0/1599916367221?e=1617235200&v=beta&t=s82wKFdMDqmEl0wD_tf5TiN5TWetevV3ZhXltG5uy4Y",
   },
   {
     quarter: "July - September 2020",
     name: "Yee Han Chung",
     position: "Nilai Leader",
     linkedin: "https://www.linkedin.com/in/cyeehan/",
-    image: "../../static/img/squad-profile-images/yee-han-chung.jpeg",
+    image: "https://media-exp1.licdn.com/dms/image/C5603AQEWJ4k9ok378A/profile-displayphoto-shrink_400_400/0/1593272507368?e=1617235200&v=beta&t=pAlVh4xvzyyK4NqM0XaYSc7GlktL_FVPd7efhHrZc0E",
   },
   {
     quarter: "July - September 2020",
     name: "Rafia Ferdous",
     position: "Subang Vice Leader",
     linkedin: "https://www.linkedin.com/in/rafia-ferdous-mim-29b015196/",
-    image: "../../static/img/squad-profile-images/rafia-ferdous.jpg",
+    image: "https://media-exp1.licdn.com/dms/image/C5103AQHJ7X8f5txozw/profile-displayphoto-shrink_400_400/0/1573086695470?e=1617840000&v=beta&t=eWGDlQzN5G1dVoi9oYqRXnyV2ZhEMkaCS5D9IdxJvKI",
   },
   {
     quarter: "July - September 2020",
     name: "Bee Yean Lim",
     position: "Nilai Vice Leader",
     linkedin: "https://www.linkedin.com/in/lim-bee-yean-002912151/",
-    image: "../../static/img/squad-profile-images/bee-yean-lim.jpg",
+    image: "https://media-exp1.licdn.com/dms/image/C5603AQFnHGLUXOGpZA/profile-displayphoto-shrink_400_400/0/1599054878487?e=1617840000&v=beta&t=OVsJbfkLXfZ2XAstU15UtL1aABsXdZothnh8zaxaQi4",
   },
   {
     quarter: "April - June 2020",
     name: "Kai Lun Ho",
     position: "Subang Leader",
     linkedin: "https://www.linkedin.com/in/hokailun/",
-    image: "../../static/img/squad-profile-images/kai-lun-ho.jpg",
+    image: "https://media-exp1.licdn.com/dms/image/C5603AQELEfJ-doj2AA/profile-displayphoto-shrink_400_400/0/1592283936109?e=1617840000&v=beta&t=Ir7yhC6O2_C8lDd7ckWpgi9fV1JHaH5RAsw2r0Ly2CE",
   },
   {
     quarter: "April - June 2020",
     name: "Kashmiethaa V. Chandran",
     position: "Nilai Leader",
     linkedin: "https://www.linkedin.com/in/kashmiethaa-v-chandran-319970182/",
-    image: "../../static/img/squad-profile-images/kashmiethaa-v-chandran.jpg",
+    image: "https://media-exp1.licdn.com/dms/image/C5103AQFagObRbVuL0g/profile-displayphoto-shrink_400_400/0/1573790488418?e=1617840000&v=beta&t=N2Y60gLcPIzOtbkzh-X2eAtkFfADbUGdwk-A7UeaCd4",
   },
   {
     quarter: "April - June 2020",
     name: "Yi Wen Foong",
     position: "Subang CFO",
     linkedin: "https://www.linkedin.com/in/yi-wen-foong/",
-    image: "../../static/img/squad-profile-images/yi-wen-foong.jpg",
+    image: "https://media-exp1.licdn.com/dms/image/C5103AQFvucxtEaOYEQ/profile-displayphoto-shrink_400_400/0/1582861242353?e=1617840000&v=beta&t=PpjdD0eaQI9Tg3B2daVcDDw3bEmJ01_tfSM899f5qWI",
   },
   {
     quarter: "April - June 2020",
     name: "Gladys Loh",
     position: "Nilai CFO",
     linkedin: "https://www.linkedin.com/in/gladysloh/",
-    image: "../../static/img/squad-profile-images/gladys-loh.jpg",
+    image: "https://media-exp1.licdn.com/dms/image/C5103AQEUrTL2Jeo2Xw/profile-displayphoto-shrink_400_400/0/1583323858549?e=1617840000&v=beta&t=UoDAeOHNNrKVTNnlFToc-r1P3mT18-9TiK2yFzqZGL4",
   },
   {
     quarter: "January - March 2020",
     name: "Sze Yin Cheah (Evelyn)",
     position: "Subang Leader",
     linkedin: "https://www.linkedin.com/in/szeyincheahevelyn/",
-    image: "../../static/img/squad-profile-images/sze-yin-cheah-evelyn.jpg",
+    image: "https://media-exp1.licdn.com/dms/image/C5103AQGgtaVBxrvNeQ/profile-displayphoto-shrink_400_400/0/1582187631816?e=1617235200&v=beta&t=fTX8NsGA0j2tFmVh1bGFzM38EDRhvNh63hdEMZu19K4",
   },
   {
     quarter: "January - March 2020",
     name: "Michelle Machingaidze",
     position: "Nilai Leader",
     linkedin: "https://www.linkedin.com/in/michelle-machingaidze/",
-    image: "../../static/img/squad-profile-images/michelle-machingaidze.jpg",
+    image: "https://media-exp1.licdn.com/dms/image/C5103AQGKJxaQDBr5Ww/profile-displayphoto-shrink_100_100/0/1581693345068?e=1617840000&v=beta&t=GLFASl5sMBqmCdIQlrNE7mIdRACuo7zJUS3D2MwfihA",
   },
   {
     quarter: "January - March 2020",
     name: "Lisa Pui Sze Teng",
     position: "Subang CFO",
     linkedin: "https://www.linkedin.com/in/lisa-pui-sze-teng-728117173/",
-    image: "../../static/img/squad-profile-images/lisa-pui-sze-teng.jpg",
+    image: "https://media-exp1.licdn.com/dms/image/C5103AQE3CIr5K3zhVw/profile-displayphoto-shrink_400_400/0/1563027975781?e=1617840000&v=beta&t=WtZd4FpH7kr7v2Lu3agMwuFvIFPKsjSW3qRarmQ7mc0",
   },
   {
     quarter: "January - March 2020",
     name: "Jun Wen Chin",
     position: "Nilai CFO",
     linkedin: "https://www.linkedin.com/in/junwen888/",
-    image: "../../static/img/squad-profile-images/jun-wen-chin.jpg",
+    image: "https://media-exp1.licdn.com/dms/image/C5103AQFGnOe13KDscw/profile-displayphoto-shrink_400_400/0/1581924795482?e=1617840000&v=beta&t=HeepJy9iYPxyCLA259syO2xD2ELI989wQ6DbqmvA7JM",
   },
   {
     quarter: "October - December 2019",
     name: "Amanda Kar Kei Aw",
     position: "Subang Leader",
     linkedin: "https://www.linkedin.com/in/awkarkei/",
-    image: "../../static/img/squad-profile-images/amanda-kar-kei-aw.jpg",
+    image: "https://media-exp1.licdn.com/dms/image/C5103AQHvmcK2cgEokQ/profile-displayphoto-shrink_400_400/0/1574666393481?e=1617840000&v=beta&t=SL4JFqe2MQ2eWvi6UzQ0N6F9XSXnNbEZjg2cUoCXMzU",
   },
   {
     quarter: "October - December 2019",
     name: "Praveendira Rao Madhavan",
     position: "Nilai Leader",
     linkedin: "https://www.linkedin.com/in/praveendira-rao-m-3683aaa1/",
-    image: "../../static/img/squad-profile-images/praveendira-rao-madhavan.jpg",
+    image: "https://media-exp1.licdn.com/dms/image/C5603AQGApVkr6ovcgQ/profile-displayphoto-shrink_400_400/0/1602762714155?e=1617840000&v=beta&t=BHMRdlHvUH1xmObD-qzXoea86wtiqrjAYNgdz5pNnzU",
   },
   {
     quarter: "October - December 2019",
     name: "Lisa Pui Sze Teng",
     position: "Subang CFO",
     linkedin: "https://www.linkedin.com/in/lisa-pui-sze-teng-728117173/",
-    image: "../../static/img/squad-profile-images/lisa-pui-sze-teng.jpg",
+    image: "https://media-exp1.licdn.com/dms/image/C5103AQE3CIr5K3zhVw/profile-displayphoto-shrink_400_400/0/1563027975781?e=1617840000&v=beta&t=WtZd4FpH7kr7v2Lu3agMwuFvIFPKsjSW3qRarmQ7mc0",
   },
   {
     quarter: "October - December 2019",
     name: "Jia Min Ong",
     position: "Nilai CFO",
     linkedin: "https://www.linkedin.com/in/jia-min-ong-a2b3a2170/",
-    image: "../../static/img/squad-profile-images/jia-min-ong.jpg",
+    image: "https://media-exp1.licdn.com/dms/image/C5103AQFQ1UFPo6QdtA/profile-displayphoto-shrink_400_400/0/1539152624037?e=1617840000&v=beta&t=Do7Ufg1DpAQWpoqgSCUorT52MzHkfX8gG9XEe3qef24",
   },
 ];
