+++ website/contents/data/squad-member.js
@@ -1,0 +1,137 @@
+export const squadMembers = [
+  {
+    name: "Ain Batrisyia",
+    linkedin: "https://www.linkedin.com/in/ainbatrisyiashaharuddin/",
+    image: "https://media-exp1.licdn.com/dms/image/C5603AQHUqTLqDkCzZg/profile-displayphoto-shrink_400_400/0/1611857065347?e=1617235200&v=beta&t=XgAJ830wrVMth4qK-6xyQOhNM4AsccyDCH-WVWd-G_8",
+  },
+  {
+    name: "Amanda Kar Kei Aw",
+    linkedin: "https://www.linkedin.com/in/awkarkei/",
+    image: "https://media-exp1.licdn.com/dms/image/C5103AQHvmcK2cgEokQ/profile-displayphoto-shrink_400_400/0/1574666393481?e=1617840000&v=beta&t=SL4JFqe2MQ2eWvi6UzQ0N6F9XSXnNbEZjg2cUoCXMzU",
+  },
+  {
+    name: "Bee Yean Lim",
+    linkedin: "https://www.linkedin.com/in/lim-bee-yean-002912151/",
+    image: "https://media-exp1.licdn.com/dms/image/C5603AQFnHGLUXOGpZA/profile-displayphoto-shrink_400_400/0/1599054878487?e=1617840000&v=beta&t=OVsJbfkLXfZ2XAstU15UtL1aABsXdZothnh8zaxaQi4",
+  },
+  {
+    name: "Gladys Loh",
+    linkedin: "https://www.linkedin.com/in/gladysloh/",
+    image: "https://media-exp1.licdn.com/dms/image/C5103AQEUrTL2Jeo2Xw/profile-displayphoto-shrink_400_400/0/1583323858549?e=1617840000&v=beta&t=UoDAeOHNNrKVTNnlFToc-r1P3mT18-9TiK2yFzqZGL4",
+  },
+  {
+    name: "Hira Khalique",
+    linkedin: "https://www.linkedin.com/in/hira-khalique-9871211a3/",
+    image: "https://media-exp1.licdn.com/dms/image/C5603AQG0FeISTDq8Ww/profile-displayphoto-shrink_400_400/0/1601366509956?e=1617235200&v=beta&t=1AvLcUYau37zfrUX4MgdY2JzB67D3Z70fN7ELbZXG78",
+  },
+  {
+    name: "Jernyn Kah Yee Ho",
+    linkedin: "https://www.linkedin.com/in/jernyn-ho-b1a179141/",
+    image: "https://media-exp1.licdn.com/dms/image/C5603AQFLdf4e9Iop1A/profile-displayphoto-shrink_400_400/0/1516963714518?e=1617840000&v=beta&t=0ORssVnh_2t1iHkMtFzJT9ooEqshOoLyFz0LJqbtCOg",
+  },
+  {
+    name: "Jia Min Ong",
+    linkedin: "https://www.linkedin.com/in/jia-min-ong-a2b3a2170/",
+    image: "https://media-exp1.licdn.com/dms/image/C5103AQFQ1UFPo6QdtA/profile-displayphoto-shrink_400_400/0/1539152624037?e=1617840000&v=beta&t=Do7Ufg1DpAQWpoqgSCUorT52MzHkfX8gG9XEe3qef24",
+  },
+  {
+    name: "Jin En Ng",
+    linkedin: "https://www.linkedin.com/in/jin-en-ng-a343b51a3/",
+    image: "https://media-exp1.licdn.com/dms/image/C5103AQGZcxEEp7aKYg/profile-displayphoto-shrink_400_400/0/1582458683616?e=1617235200&v=beta&t=iX3x0h_ID5yAkXiXvP5yYV9WAQDSn3j-EfYuJuecWKI",
+  },
+  {
+    name: "Joanne Lou",
+    linkedin: "https://www.linkedin.com/in/joanne-lou-54b10b186/",
+    image: "https://media-exp1.licdn.com/dms/image/C5603AQGbjZSj1c19Xg/profile-displayphoto-shrink_400_400/0/1601440694562?e=1617235200&v=beta&t=FgqwJ9mmzpzMKDxcAwHG-6G2oLEvb53g-F4lKV1gf7U",
+  },
+  {
+    name: "Jun Wen Chin",
+    linkedin: "https://www.linkedin.com/in/junwen888/",
+    image: "https://media-exp1.licdn.com/dms/image/C5103AQFGnOe13KDscw/profile-displayphoto-shrink_400_400/0/1581924795482?e=1617840000&v=beta&t=HeepJy9iYPxyCLA259syO2xD2ELI989wQ6DbqmvA7JM",
+  },
+  {
+    name: "Jye Yie Tee",
+    linkedin: "https://www.linkedin.com/in/tee-jye-yie-8a48831a1/",
+    image: "https://media-exp1.licdn.com/dms/image/C5603AQFjpuAhWSmX9Q/profile-displayphoto-shrink_400_400/0/1603036447216?e=1617840000&v=beta&t=L51kzIjMKQ7moeyNSrAaZZCQ0BuYAyUENLka6kOiY1g",
+  },
+  {
+    name: "Kai Lun Ho",
+    linkedin: "https://www.linkedin.com/in/hokailun/",
+    image: "https://media-exp1.licdn.com/dms/image/C5603AQELEfJ-doj2AA/profile-displayphoto-shrink_400_400/0/1592283936109?e=1617840000&v=beta&t=Ir7yhC6O2_C8lDd7ckWpgi9fV1JHaH5RAsw2r0Ly2CE",
+  },
+  {
+    name: "Kashmiethaa V. Chandran",
+    linkedin: "https://www.linkedin.com/in/kashmiethaa-v-chandran-319970182/",
+    image: "https://media-exp1.licdn.com/dms/image/C5103AQFagObRbVuL0g/profile-displayphoto-shrink_400_400/0/1573790488418?e=1617840000&v=beta&t=N2Y60gLcPIzOtbkzh-X2eAtkFfADbUGdwk-A7UeaCd4",
+  },
+  {
+    name: "Krena Shree",
+    linkedin: "https://www.linkedin.com/in/krena-shree-0953b61a3/",
+    image: "https://media-exp1.licdn.com/dms/image/C5603AQHJbZm85ItvdQ/profile-displayphoto-shrink_400_400/0/1611498868534?e=1617235200&v=beta&t=AXegKmg6SwkH3OUSo8kcBitrBy7bnwXH5sE6qSGSjUQ",
+  },
+  {
+    name: "Lisa Pui Sze Teng",
+    linkedin: "https://www.linkedin.com/in/lisa-pui-sze-teng-728117173/",
+    image: "https://media-exp1.licdn.com/dms/image/C5103AQE3CIr5K3zhVw/profile-displayphoto-shrink_400_400/0/1563027975781?e=1617840000&v=beta&t=WtZd4FpH7kr7v2Lu3agMwuFvIFPKsjSW3qRarmQ7mc0",
+  },
+  {
+    name: "Michelle Machingaidze",
+    linkedin: "https://www.linkedin.com/in/michelle-machingaidze/",
+    image: "https://media-exp1.licdn.com/dms/image/C5103AQGKJxaQDBr5Ww/profile-displayphoto-shrink_100_100/0/1581693345068?e=1617840000&v=beta&t=GLFASl5sMBqmCdIQlrNE7mIdRACuo7zJUS3D2MwfihA",
+  },
+  {
+    name: "Mun Hong Teoh",
+    linkedin: "https://www.linkedin.com/in/teoh-mun-hong-442190bb/",
+    image: "https://media-exp1.licdn.com/dms/image/C5103AQFp9BwTcZeSgg/profile-displayphoto-shrink_400_400/0/1587051911092?e=1617235200&v=beta&t=QG9mYVvPpV4O7OfDfsm9WQYsYhsVaW-cJi-hgjIMGR4",
+  },
+  {
+    name: "Phoenix Yee",
+    linkedin: "https://www.linkedin.com/in/phoenixyee/",
+    image: "https://media-exp1.licdn.com/dms/image/C5103AQH5cZqnsP_M1w/profile-displayphoto-shrink_400_400/0/1524567457288?e=1617840000&v=beta&t=IEqh0az-Yd5l-Eo0YPixHRNR8g6b2v3AUVVrILoxWhk",
+  },
+  {
+    name: "Poh Yee Lew",
+    linkedin: "https://www.linkedin.com/in/poh-yee-lew-a0b4671a2/",
+    image: "https://media-exp1.licdn.com/dms/image/C5103AQFLNCs725TAxQ/profile-displayphoto-shrink_400_400/0/1581926433402?e=1617840000&v=beta&t=iaHXuTzixpnANDV86Iy6Utop2JWDr0xTnhRfC_zoDGo",
+  },
+  {
+    name: "Praveendira Rao Madhavan",
+    linkedin: "https://www.linkedin.com/in/praveendira-rao-m-3683aaa1/",
+    image: "https://media-exp1.licdn.com/dms/image/C5603AQGApVkr6ovcgQ/profile-displayphoto-shrink_400_400/0/1602762714155?e=1617840000&v=beta&t=BHMRdlHvUH1xmObD-qzXoea86wtiqrjAYNgdz5pNnzU",
+  },
+  {
+    name: "Rafia Ferdous",
+    linkedin: "https://www.linkedin.com/in/rafia-ferdous-mim-29b015196/",
+    image: "https://media-exp1.licdn.com/dms/image/C5103AQHJ7X8f5txozw/profile-displayphoto-shrink_400_400/0/1573086695470?e=1617840000&v=beta&t=eWGDlQzN5G1dVoi9oYqRXnyV2ZhEMkaCS5D9IdxJvKI",
+  },
+  {
+    name: "Stephanie Chong",
+    linkedin: "https://www.linkedin.com/in/stephanie-chong-7b36291a4/",
+    image: "https://media-exp1.licdn.com/dms/image/C5603AQEwqOq5NM3kfA/profile-displayphoto-shrink_400_400/0/1611306039392?e=1617235200&v=beta&t=AnUmsC82ZX8kyUfvA_oWS1DyCJaRRH8MyKKT8rE0FMI",
+  },
+  {
+    name: "Sze Yin Cheah (Evelyn)",
+    linkedin: "https://www.linkedin.com/in/szeyincheahevelyn/",
+    image: "https://media-exp1.licdn.com/dms/image/C5103AQGgtaVBxrvNeQ/profile-displayphoto-shrink_400_400/0/1582187631816?e=1617235200&v=beta&t=fTX8NsGA0j2tFmVh1bGFzM38EDRhvNh63hdEMZu19K4",
+  },
+  {
+    name: "Yee Han Chung",
+    linkedin: "https://www.linkedin.com/in/cyeehan/",
+    image: "https://media-exp1.licdn.com/dms/image/C5603AQEWJ4k9ok378A/profile-displayphoto-shrink_400_400/0/1593272507368?e=1617235200&v=beta&t=pAlVh4xvzyyK4NqM0XaYSc7GlktL_FVPd7efhHrZc0E",
+  },
+  {
+    name: "Yi Wen Foong",
+    linkedin: "https://www.linkedin.com/in/yi-wen-foong/",
+    image: "https://media-exp1.licdn.com/dms/image/C5103AQFvucxtEaOYEQ/profile-displayphoto-shrink_400_400/0/1582861242353?e=1617840000&v=beta&t=PpjdD0eaQI9Tg3B2daVcDDw3bEmJ01_tfSM899f5qWI",
+  },
+  {
+    name: "Zhen Hao Lee",
+    linkedin: "https://www.linkedin.com/in/lee-zhen-hao-a28a96155/",
+    image: "https://media-exp1.licdn.com/dms/image/C5603AQEA5T3pKE6mFw/profile-displayphoto-shrink_400_400/0/1599916367221?e=1617235200&v=beta&t=s82wKFdMDqmEl0wD_tf5TiN5TWetevV3ZhXltG5uy4Y",
+  },
+  {
+    name: "Zi Ching Goh",
+    linkedin: "https://www.linkedin.com/in/zi-ching-g-750b26198/",
+    image: "https://media-exp1.licdn.com/dms/image/C5103AQEFXWZ97GxaNg/profile-displayphoto-shrink_400_400/0/1574999230863?e=1617235200&v=beta&t=axsNPLoEvA5xeYV0qqqw_GXD1uZw8Oi4BpnjgsNlqWU",
+  },
+];
