--- website/contents/data/squad-members.js
@@ -1,225 +1,0 @@
-export const squadMembers = [
-  {
-    name: "Ainaa Radhiah Binti Nur Azam",
-    linkedin: "https://www.linkedin.com/in/ainaarna",
-    image: "../../static/img/squad-profile-images/ainaa-azam.png",
-  },
-  {
-    name: "Ain Batrisyia",
-    linkedin: "https://www.linkedin.com/in/ainbatrisyiashaharuddin/",
-    image: "../../static/img/squad-profile-images/ain-batrisyia.jpeg",
-  },
-  {
-    name: "Amanda Kar Kei Aw",
-    linkedin: "https://www.linkedin.com/in/awkarkei/",
-    image: "../../static/img/squad-profile-images/amanda-kar-kei-aw.jpg",
-  },
-  {
-    name: "Ariful Islam",
-    linkedin: "https://www.linkedin.com/in/arif-ai/",
-    image: "../../static/img/squad-profile-images/ariful-islam.jpg",
-  },
-  {
-    name: "Bee Yean Lim",
-    linkedin: "https://www.linkedin.com/in/lim-bee-yean-002912151/",
-    image: "../../static/img/squad-profile-images/bee-yean-lim.jpg",
-  },
-  {
-    name: "Cze Yik Lee",
-    linkedin: "https://www.linkedin.com/in/cze-yik-lee-980a041a9/",
-    image: "../../static/img/squad-profile-images/cze-yik-lee.jpg",
-  },
-  /*
-  {
-    name: "Dafina Siti Marsya Amanda",
-    linkedin: "----------------------------",
-    image: "-------------------------",
-  },
-  */
-  {
-    name: "Fadhilah Nur Rania",
-    linkedin: "https://www.linkedin.com/in/fnrania/",
-    image: "../../static/img/squad-profile-images/fadhilah-nur-rania.png",
-  },
-  {
-    name: "Fui May Chong",
-    linkedin: "https://www.linkedin.com/in/chong-fui-may-421631204",
-    image: "../../static/img/squad-profile-images/fui-may-chong.png",
-  },
-  /*
-  {
-    name: "Fu Shin Hui",
-    linkedin: "----------------------------",
-    image: "-------------------------",
-  },  
-  */
-  {
-    name: "Gladys Loh",
-    linkedin: "https://www.linkedin.com/in/gladysloh/",
-    image: "../../static/img/squad-profile-images/gladys-loh.jpg",
-  },
-  {
-    name: "Hira Khalique",
-    linkedin: "https://www.linkedin.com/in/hira-khalique-9871211a3/",
-    image: "../../static/img/squad-profile-images/hira-khalique.jpg",
-  },
-  {
-    name: "Isabella Chee",
-    linkedin: "https://www.linkedin.com/in/isabella-chee-3142791b6/",
-    image: "../../static/img/squad-profile-images/isabella-chee.png",
-  },
-  {
-    name: "Jernyn Kah Yee Ho",
-    linkedin: "https://www.linkedin.com/in/jernyn-ho-b1a179141/",
-    image: "../../static/img/squad-profile-images/jernyn-kah-yee-ho.jpg",
-  },
-  {
-    name: "Jia Min Ong",
-    linkedin: "https://www.linkedin.com/in/jia-min-ong-a2b3a2170/",
-    image: "../../static/img/squad-profile-images/jia-min-ong.jpg",
-  },
-  {
-    name: "Jin En Ng",
-    linkedin: "https://www.linkedin.com/in/jin-en-ng-a343b51a3/",
-    image: "../../static/img/squad-profile-images/jin-en-ng.jpg",
-  },
-  {
-    name: "Joanne Lou",
-    linkedin: "https://www.linkedin.com/in/joanne-lou-54b10b186/",
-    image: "../../static/img/squad-profile-images/joanne-lou.jpg",
-  },
-  {
-    name: "Jun Wen Chin",
-    linkedin: "https://www.linkedin.com/in/junwen888/",
-    image: "../../static/img/squad-profile-images/jun-wen-chin.jpg",
-  },
-  {
-    name: "Jye Yie Tee",
-    linkedin: "https://www.linkedin.com/in/tee-jye-yie-8a48831a1/",
-    image: "../../static/img/squad-profile-images/jye-yie-tee.jpg",
-  },
-  {
-    name: "Kai Lun Ho",
-    linkedin: "https://www.linkedin.com/in/hokailun/",
-    image: "../../static/img/squad-profile-images/kai-lun-ho.jpg",
-  },
-  {
-    name: "Kashmiethaa V. Chandran",
-    linkedin: "https://www.linkedin.com/in/kashmiethaa-v-chandran-319970182/",
-    image: "../../static/img/squad-profile-images/kashmiethaa-v-chandran.jpg",
-  },
-  {
-    name: "Kreepa Shree",
-    linkedin: "https://www.linkedin.com/in/kreepa-shree-paidathalli-5a48a6204",
-    image: "../../static/img/squad-profile-images/kreepa-shree-ap-paidathalli.png",
-  },
-  {
-    name: "Krena Shree",
-    linkedin: "https://www.linkedin.com/in/krena-shree-0953b61a3/",
-    image: "../../static/img/squad-profile-images/krena-shree.jpg",
-  },
-  {
-    name: "Krishna Priya Udayakumar",
-    linkedin: "https://www.linkedin.com/in/krishna-priya-udayakumar-02bb76204",
-    image: "../../static/img/squad-profile-images/krishna-priya-udayakumar.png",
-  },
-  {
-    name: "Lisa Pui Sze Teng",
-    linkedin: "https://www.linkedin.com/in/lisa-pui-sze-teng-728117173/",
-    image: "../../static/img/squad-profile-images/lisa-pui-sze-teng.jpg",
-  },
-  /*
-  {
-    name: "Louis Ezra Endro",
-    linkedin: "-----------------------------------",
-    image: "----------------------------",
-  },
-  */
-  {
-    name: "Michelle Loo Xuan Wen",
-    linkedin: "https://www.linkedin.com/in/michelle-loo-xuan-wen-246874201",
-    image: "../../static/img/squad-profile-images/michelle-xuan-wen-loo.png",
-  },
-  {
-    name: "Michelle Machingaidze",
-    linkedin: "https://www.linkedin.com/in/michelle-machingaidze/",
-    image: "../../static/img/squad-profile-images/michelle-machingaidze.jpg",
-  },
-  {
-    name: "Mun Hong Teoh",
-    linkedin: "https://www.linkedin.com/in/teoh-mun-hong-442190bb/",
-    image: "../../static/img/squad-profile-images/mun-hong-teoh.jpg",
-  },
-  {
-    name: "Phoenix Yee",
-    linkedin: "https://www.linkedin.com/in/phoenixyee/",
-    image: "../../static/img/squad-profile-images/phoenix-yee.jpg",
-  },
-  {
-    name: "Poh Yee Lew",
-    linkedin: "https://www.linkedin.com/in/poh-yee-lew-a0b4671a2/",
-    image: "../../static/img/squad-profile-images/poh-yee-lew.jpg",
-  },
-  {
-    name: "Praveendira Rao Madhavan",
-    linkedin: "https://www.linkedin.com/in/praveendira-rao-m-3683aaa1/",
-    image: "../../static/img/squad-profile-images/praveendira-rao-madhavan.jpg",
-  },
-  {
-    name: "Rafia Ferdous",
-    linkedin: "https://www.linkedin.com/in/rafia-ferdous-mim-29b015196/",
-    image: "../../static/img/squad-profile-images/rafia-ferdous.jpg",
-  },
-  {
-    name: "Sher Ree Koh",
-    linkedin: "https://www.linkedin.com/in/koh-sher-ree-b8a62a204/",
-    image: "../../static/img/squad-profile-images/sher-ree-koh.png",
-  },
-  {
-    name: "Stephanie Chong",
-    linkedin: "https://www.linkedin.com/in/stephanie-chong-7b36291a4/",
-    image: "../../static/img/squad-profile-images/stephanie-chong.jpg",
-  },
-  {
-    name: "Sze Yin Cheah (Evelyn)",
-    linkedin: "https://www.linkedin.com/in/szeyincheahevelyn/",
-    image: "../../static/img/squad-profile-images/sze-yin-cheah-evelyn.jpg",
-  },
-  {
-    name: "Tuan Sihan",
-    linkedin: "https://www.linkedin.com/in/tuansihan/",
-    image: "../../static/img/squad-profile-images/tuan-sihan.png",
-  },
-  /*
-  {
-    name: "Veishnawi A/P Nehru",
-    linkedin: "-----------------",
-    image: "--------------------",
-  },
-  */
-  {
-    name: "Yee Han Chung",
-    linkedin: "https://www.linkedin.com/in/cyeehan/",
-    image: "../../static/img/squad-profile-images/yee-han-chung.jpeg",
-  },
-  {
-    name: "Yen Yoo Cheu",
-    linkedin: "https://www.linkedin.com/in/cheu-yen-yoo-4713731a4",
-    image: "../../static/img/squad-profile-images/yen-yoo-cheu.png",
-  },
-  {
-    name: "Yi Wen Foong",
-    linkedin: "https://www.linkedin.com/in/yi-wen-foong/",
-    image: "../../static/img/squad-profile-images/yi-wen-foong.jpg",
-  },
-  {
-    name: "Zhen Hao Lee",
-    linkedin: "https://www.linkedin.com/in/lee-zhen-hao-a28a96155/",
-    image: "../../static/img/squad-profile-images/zhen-hao-lee.jpg",
-  },
-  {
-    name: "Zi Ching Goh",
-    linkedin: "https://www.linkedin.com/in/zi-ching-g-750b26198/",
-    image: "../../static/img/squad-profile-images/zi-ching-goh.jpg",
-  },
-];
